Automatic cleanup of old tasks.
Tasks that have been untouched for a configurable number of days
(default: 14) will be deleted. They will be re-downloaded from the
Server if there is a server-side update.

NOTE: this requires the active database to be changed a bit, renaming
the "_updated" field on flamenco_task documents to "last_updated". This
clarifies that it's a different field from the Server-side _updated field.