Added filesystem watcher for keeping an eye on rendered images.
Uses the fsnotify library to use OS-specific notifications from the file
system. Since that library doesn't support the "file closed" event, we have
to work around this. Files that haven't been written to for 5 seconds
are considered to be closed, and are pushed through a channel to notify
the rest of the world the file is there.

Currently the list of recognised file extensions is hard-coded.