Introduced path replacement variables.
This allows Clients and Workers to run on different platforms.

Note that `variables` and `path_replacement` share a namespace -- variable
names have to be unique, and cannot be used in both `variables` and
`path_replacement` sections. If this happens, Flamenco Manager will log
the offending name, and refuse to start.