Added "setup mode", and moved core functionality to "normal mode".
When "-setup" is passed as CLI arg or when the configuration file does not
exist, "setup mode" is run. In other cases "normal mode" is run.

The setup web interface is only available in "setup mode", whereas the
normal functionality of Flamenco Server is only available in "normal mode".
This makes it very explicit in which mode the server is running, and
prevents Workers from doing things on the Manager while it's being set up.

Also switched to our own fork of gossdp. I've changed our fork so that
the logger is compatible with logrus.