Added ability to send workers to sleep (and wake them up again)
When a status change is requested, /may-i-run/{task-id} and /task will
refuse the worker to run the task or get a new one. The newly requested
status is returned instead. This keeps happening until the worker
acknowledges the new status by a call to /ack-status-change/{ack-status}.

A sleeping worker can still get a task; there is no status checks on that.
The worker will just be marked as active then. It is up to the worker to
adhere to the status given by the Manager.

The worker can poll /status-change while sleeping to know when to wake up.
In a later version we possibly could turn this into a long poll.