Refuse to let the Worker set a task status to 'active'
The scheduler sets the task to 'active' when it is assigned to a worker.
This should be the only code that activates a task; previously the Worker
was allowed to do this, which means that returning a task to 'queued' and
then flushing old updates from a Worker could inadvertently reactivate the
task and cause a timeout.