Schedule is now event-based and can be temporarily overridden
The schedule now only works at the schedule edges (so start/end of sleep
time) instead of all the time. As a result, the worker will still sleep
and wake up on the scheduled times, but can still be sent to sleep/wake
manually without influencing the scheduler.

(e.g. previously manually setting the status would deactivate the schedule;
this no longer happens).