Fix T50981 Worker deallocation from job if fails n tasks
When a Worker notifies the Manager a task failed, the number of failed
tasks of this worker, on this job, of the same task type as the
currently failed task is counted. If this count is above a threshold,
the (worker ID, job ID, task type) tuple is added to the blacklist. This
prevents the worker from getting such tasks. Matching failed tasks are
re-queued so that they can be executed by another worker.

This requies a new setting `blacklist_threshold`, which indicates the
number of failed tasks at which the above behaviour is triggered. It
defaults to 3. This means that it's likely that we should also increase
the TASK_FAIL_JOB_PERCENTAGE constant in Flamenco Server so that it's
more lenient towards failure (as excessive failure will trigger
requeueing anyway).

Note that there is NO starvation detection. In other words, if a job has
certain tasks that were failed by all available workers (and thus all
workers are blacklisted for this job & task type) there is no detection
that this happened. As a result, the job will be stuck in 'active'
status without it ever having a chance of being finished.