Upload task log files when requested from the Flamenco Server
The server can pass us (job ID, task ID) tuples in the response of the
'task-update-batch' endpoint. These tuples are then used to find the
task's log file, compress it, and send it to the Flamenco Manager.

The queue of logfiles to send is maintained by the Server. This means
we'll repeatedly get the same (job ID, task ID) until we've actually
uploaded the logfile to the Server's satisfaction. As a result, we don't
persist the requested IDs, but rely on the server to pass us the list
again if need be.