Soft-fail tasks when there are workers left to retry it
When a Worker sends a task update with `status='failed'`, that status is
actually overridden by the Manager to `status='soft-failed'` if there is
a worker that is *not* blacklisted for that specific task type/job. This
happens until the soft-failing worker is actually blacklisted, in which
case it is assumed to be an issue with the worker. All the previously
soft-failed tasks are set to `'claimed-by-manager'` so that they can be
picked up by another worker.