Allow soft-failed tasks to be run by other workers
For this we keep track of which worker failed which task (in
`Task.FailedByWorkers`). The scheduler will not assign a worker with
tasks it failed before.

When there are no more workers left to run a task (either because of
blacklisting or because all workers have tried & failed this particular
task) the status will be 'failed', otherwise 'soft-failed'.