When serving log file, conditionally only show head + tail of the log
When the task log file is uncompressed we only show the first X and last
Y kilobytes of logging. When the log file is compressed this isn't (easily)
possible, so then the file is sent as a GZipped attachment (so forcing the
browser to download it to disk instead of loading it all in memory).

When the user agent is WGet or Curl, always the entire log is served.