Fix T61846: more efficient task update check
When the scheduler finds a task for a worker, it checks with Flamenco
Server to see if it was updated. If it is, it's re-downloaded and the
scheduler is run again. Previously this would happen for all tasks
sequentially, which is rather slow when there are many changes (for example
when a job with 1000+ tasks gets cancel-requested).

Now only 3 attempts are made to refresh a task; if all those scheduling
attempts return a task that was changed on the Server, two things happen:
  - The Task Update Pusher is kicked to force a push; this fetches a list
    of cancel-requested task IDs (and cancels those tasks).
  - The Downloader is kicked to get an update of all runnable tasks.