Auto-download JWT public keys
The public keys are downloaded in a loop with a configurable interval
(default 4h). The HTTP `Last-Modified` and `If-Modified-Since` headers
are used to minimise download overhead and to only reload the keys when
they have actually changed.

After startup and after downloading new keys the Shaman performs a
self-test of the JWT tokens. If it has a private key, it tests that it
has a matching public key; if not, it logs an error.