Protect the dashboard with JWT tokens
Upon receiving a 401 response from `/as-json`, the JS code will now GET
the `/jwt/token-urls` endpoint, which returns URLs to call to A) get the
token, and B) log in if URL A) returns 403 Forbidden.

Protecting the latest image meant that we no longer only protect XHR
calls, but also regular HTTP requests from the browser. This was easiest
to do by moving the token from the `Authorization: Bearer {token}`
header to a cookie.

I also changed endpoint /jwt/get-token to /jwt/token-urls. The endpoint
was intended to redirect to a Flamenco Server URL where the token could
be obtained (hence the name), but that didn't work well with JavaScript
when the user also needs to log in on Flamenco Server. Just returning
JSON with the two URLs ('get token' and 'login') turned out to be the
simplest approach.