Secure worker registration with a JWT signed with a pre-shared secret
Both the Workers and Manager need to have the same value for their
`worker_registration_secret` configuration setting (unless the Manager
has an empty string there, in which case registration is open to anyone).

The pre-shared secret is used by the Worker to create a short-lived JWT
token, which is then used in an `Authorization: Bearer {token}` header when
POSTing to the `/register-worker` endpoint.