Removed LICENSE BLOCK lines + moved license block to top of file
The LICENSE BLOCK lines were there because Blender does that too, but they
are being/have been removed from the Blender sources.

Apparently when there is a newline between the comment and the `package`
keyword, it's no longer interpreted as package documentation (which is why
I had the comment below the `package` keyword in the first place).