Big refactor of web setup mode (JWT security + YAML editor + Vue.js)
Web setup is now optionally secured with JWT tokens. When the Manager is
not linked to a Server, this SECURITY IS TURNED OFF. Flamenco Server is
used to supply us with JWT tokens and public keys for token validation;
without knowing which server to connect to, this workflow is impossible.

Other changes are:
- Immediately download JWT keys upon starting. This is most important
  for the setup mode, where we may need to download keys immediately
  after restarting (which happens after linking).
- Web setup now uses main layout.html file and Vue.js.
- Web setup now loads and saves settings via YAML. It can show a web
  form as well as an advanced mode that allows direct editing of YAML.

Note that the YAML is never sent byte-for-byte from the config file to
the web frontend. It is always parsed, modified (the `ManagerSecret`
property is cleared), and then re-marshalled to YAML before sending to
the frontend. The frontend does the same: it first parses the YAML from
the editor (showing a clear description about any parse errors), then
converts the parsed configuration back to YAML before sending it to the
server.