Settings version 2
Settings are now versioned; unversioned settings are implied to be
version 1. This commit introduces settings version 2, which changes the
way variables work. This requires Flamenco Server > 2.4.

- Path replacement and regular variables have been merged into one set
  of variables.
- Path replacement variables are now marked with `direction: twoway`,
  whereas what used to be regular variables are now marked with
  `direction: oneway`.
- Values can now have an audience, which is either 'all' (the old
  behaviour and implied when `audience=''`), 'workers' (only sent to
  workers), or 'users' (sent to Flamenco Server so that users see them
  in the web interface + Blender Cloud add-on).

Variables are automatically converted to version 2 when loading the
settings, even when the settings are still specifying the old style.
When saving the configuration file it is saved as version 2.