Allow hiding infrastructure-dependent settings in web setup
When `websetup.hide_infra_settings` is true, the websetup will hide
certain settings that are infrastructure-specific. For example, it hides
MongoDB choice, port numbers, task log directory, all kind of
infra/hosting-specific things. This is used, for example, by the
automated Azure deployment to avoid messing up settings that are
specific to that particular installation.