Web setup: Start of work on predefined variables
To make it simpler for people to work with variables, I think it's a
good idea to create some predefined ones (and not them be deleted). This
also allows us to include some indication as to the semantics of each
variable in the web setup, making it easier to work with.